requirements.txt
faster-whisper>=1.0.0
torch>=2.0.0
python-socketio[client]>=5.16.0
sounddevice>=0.5.0
numpy>=1.24.0
scipy>=1.10.0
huggingface-hub>=0.19.0
python-dotenv>=1.0.0
